srcs/back/user_mgmt/requirements.txt
Django==4.2.7
channels~=3.0
bcrypt
django-cors-headers
django-cors-middleware
psycopg2
djangorestframework
